Refactored App component into functional component using Hooks for slightly better performance and moved header related constants to Header component
